# packages/contract/openapi.yaml
openapi: 3.1.0
info:
  title: RMC ERP API
  version: 0.2.0
  description: >
    Contract-first source of truth. The TS client, React Query hooks, and
    Zod request/response schemas consumed by apps/web and apps/mobile are
    generated from this file — see CLAUDE.md. Do not hand-write a fetch
    call that bypasses the generated hooks.

    Money fields are JOD decimal strings (e.g. "1234.567"), never numbers
    — formatting/parsing to the internal integer-fils representation
    happens at this HTTP boundary (packages/core's jodStringToFils /
    filsToJodString), matching CLAUDE.md's money conventions.
servers:
  - url: /api
...
paths:
...
  /price-lists:
    get:
      operationId: listPriceLists
      tags: [priceLists]
      summary: List price lists.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/QParam"
        - $ref: "#/components/parameters/FormatParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PriceListListResponse" }
            text/csv:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createPriceList
      tags: [priceLists]
      summary: Create a price list (exactly one of projectId/customerId/branchId must match its tier).
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePriceListRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PriceList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "400": { $ref: "#/components/responses/ValidationError" }
  /price-lists/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    get:
      operationId: getPriceList
      tags: [priceLists]
      summary: Get a price list (with its lines) by id.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PriceListWithLines" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      operationId: updatePriceList
      tags: [priceLists]
      summary: Update a price list's name/active state.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePriceListRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PriceList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "400": { $ref: "#/components/responses/ValidationError" }
    delete:
      operationId: voidPriceList
      tags: [priceLists]
      summary: Soft-delete (void) a price list.
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VoidRequest" }
      responses:
        "204": { description: Voided }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /price-lists/{priceListId}/lines:
    parameters:
      - name: priceListId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: createPriceListLine
      tags: [priceLists]
      summary: Add a product price line to a price list.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePriceListLineRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PriceListLine" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "400": { $ref: "#/components/responses/ValidationError" }
  /price-lists/{priceListId}/lines/{id}:
    parameters:
      - name: priceListId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - $ref: "#/components/parameters/IdParam"
    put:
      operationId: updatePriceListLine
      tags: [priceLists]
      summary: Update a price list line.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePriceListLineRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PriceListLine" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "400": { $ref: "#/components/responses/ValidationError" }
    delete:
      operationId: voidPriceListLine
      tags: [priceLists]
      summary: Soft-delete (void) a price list line.
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VoidRequest" }
      responses:
        "204": { description: Voided }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1 }
    PageSizeParam:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100 }
    QParam:
      name: q
      in: query
      description: Free-text search.
      schema: { type: string }
    FormatParam:
      name: format
      in: query
      schema: { type: string, enum: [json, csv] }
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Missing permission
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, code]
          properties:
            message: { type: string }
            code: { type: string }
            details: { type: object }
    VoidRequest:
      type: object
      properties:
        reason: { type: string }
...
    PriceList:
      type: object
      required: [id, companyId, tier, projectId, customerId, branchId, name, isActive, createdAt, updatedAt, voidedAt]
      properties:
        id: { type: string, format: uuid }
        companyId: { type: string, format: uuid }
        tier: { type: string, enum: [project, customer, branch, company] }
        projectId: { type: string, format: uuid, nullable: true }
        customerId: { type: string, format: uuid, nullable: true }
        branchId: { type: string, format: uuid, nullable: true }
        name: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        voidedAt: { type: string, format: date-time, nullable: true }
    CreatePriceListRequest:
      type: object
      required: [tier, name]
      properties:
        tier: { type: string, enum: [project, customer, branch, company] }
        name: { type: string, minLength: 1 }
        projectId: { type: string, format: uuid, nullable: true }
        customerId: { type: string, format: uuid, nullable: true }
        branchId: { type: string, format: uuid, nullable: true }
        isActive: { type: boolean }
    UpdatePriceListRequest:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        isActive: { type: boolean }
    PriceListListResponse:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/PriceList" } }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    PriceListLine:
      type: object
      required:
        [id, companyId, priceListId, productId, concreteUnitPriceJod, deliveryUnitPriceJod,
         taxRateBasisPoints, effectiveFrom, effectiveTo, createdAt, updatedAt, voidedAt]
      properties:
        id: { type: string, format: uuid }
        companyId: { type: string, format: uuid }
        priceListId: { type: string, format: uuid }
        productId: { type: string, format: uuid }
        concreteUnitPriceJod: { type: string }
        deliveryUnitPriceJod: { type: string }
        taxRateBasisPoints: { type: integer }
        effectiveFrom: { type: string, format: date-time }
        effectiveTo: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        voidedAt: { type: string, format: date-time, nullable: true }
    CreatePriceListLineRequest:
      type: object
      required: [productId, concreteUnitPriceJod, deliveryUnitPriceJod, effectiveFrom]
      properties:
        productId: { type: string, format: uuid }
        concreteUnitPriceJod: { type: string }
        deliveryUnitPriceJod: { type: string }
        taxRateBasisPoints: { type: integer }
        effectiveFrom: { type: string, format: date-time }
        effectiveTo: { type: string, format: date-time, nullable: true }
    UpdatePriceListLineRequest:
      type: object
      properties:
        concreteUnitPriceJod: { type: string }
        deliveryUnitPriceJod: { type: string }
        taxRateBasisPoints: { type: integer }
        effectiveFrom: { type: string, format: date-time }
        effectiveTo: { type: string, format: date-time, nullable: true }
    PriceListWithLines:
      allOf:
        - $ref: "#/components/schemas/PriceList"
        - type: object
          required: [lines]
          properties:
            lines: { type: array, items: { $ref: "#/components/schemas/PriceListLine" } }
...